removed redundatn code
