+++ src/index.html
@@ -13,7 +13,7 @@
     <section id="hotel-modal" class="modal">
         <div class="modal-content">
             <header class="modal-header">
-                <div class="modal-title">Update Hotel 1</div>
+                <div class="modal-title" id="hotelModalTitle"></div>
                 <span class="modal-close">&times;</span>
                 
             </header>
@@ -21,7 +21,7 @@
                 <form id="hotelUpdateForm">
                     <div class="form-element">
                         <label for="hotel-name-input" class="form-label">Name: </label>
-                        <input type="text" class="form-input" name="hotel-name-input" id="hotel-name-input">
+                        <input type="text" class="form-input" name="hotelNameInput">
                     </div>
                     <div class="form-element">
                         <div class="form-label">Choose that apply:</div>
